@@ -29,15 +29,14 @@
     "@reduxjs/toolkit": "^1.8.1",
     "@types/node": "^16.11.26",
     "@types/node-sass": "^4.11.2",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
     "@types/react-router-dom": "^5.3.3",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.2.0",
+    "react": "^18.0.0",
     "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-redux": "^8.0.0",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
